Add form in style.css
Creating the tag that represents it is <form>.
The <input> tag is used for user input data.
To create a label for the input, use the <label> tag.
To connect an input with its label, we assign an id to the input and associate that id with the "for" attribute of the label.
Some types of input, such as text and submit, have a label with a display inline and the input has a display inline-block.